+++ src/main/java/politech/budget/dto/BarChart.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class BarChart {
 
     private String text;
-    private Integer credit;
-    private Integer debit;
-    private Integer amount;
+    private Double credit;
+    private Double debit;
+    private Double amount;
 }
